LayoutTests/imported/w3c/web-platform-tests/css/css-values/calc-complex-unresolved-serialize-expected.txt


FAIL 'calc(pow(2, sign(1em - 18px)))' as a specified value should serialize as 'calc(pow(2, sign(1em - 18px)))'. assert_equals: 'calc(pow(2, sign(1em - 18px)))' should round-trip exactly in specified values. expected "calc(pow(2, sign(1em - 18px)))" but got "pow(2, sign(1em - 18px))"
PASS 'calc(pow(2, sign(1em - 18px)))' as a computed value should serialize as '2'.
FAIL 'calc(pow(sign(1em - 18px), 2))' as a specified value should serialize as 'calc(pow(sign(1em - 18px), 2))'. assert_equals: 'calc(pow(sign(1em - 18px), 2))' should round-trip exactly in specified values. expected "calc(pow(sign(1em - 18px), 2))" but got "pow(sign(1em - 18px), 2)"
PASS 'calc(pow(sign(1em - 18px), 2))' as a computed value should serialize as '1'.
FAIL 'calc(pow(sign(1em - 18px), sign(1em - 18px)))' as a specified value should serialize as 'calc(pow(sign(1em - 18px), sign(1em - 18px)))'. assert_equals: 'calc(pow(sign(1em - 18px), sign(1em - 18px)))' should round-trip exactly in specified values. expected "calc(pow(sign(1em - 18px), sign(1em - 18px)))" but got "pow(sign(1em - 18px), sign(1em - 18px))"
PASS 'calc(pow(sign(1em - 18px), sign(1em - 18px)))' as a computed value should serialize as '1'.
FAIL 'calc(pow(2, sibling-index())' as a specified value should serialize as 'calc(pow(2, sibling-index()))'. assert_equals: 'calc(pow(2, sibling-index()))' should round-trip exactly in specified values. expected "calc(pow(2, sibling-index()))" but got "pow(2, sibling-index())"
PASS 'calc(pow(2, sibling-index())' as a computed value should serialize as '2'.
FAIL 'calc(pow(sibling-index(), 2)' as a specified value should serialize as 'calc(pow(sibling-index(), 2))'. assert_equals: 'calc(pow(sibling-index(), 2))' should round-trip exactly in specified values. expected "calc(pow(sibling-index(), 2))" but got "pow(sibling-index(), 2)"
PASS 'calc(pow(sibling-index(), 2)' as a computed value should serialize as '1'.
FAIL 'calc(pow(sibling-index(), sibling-index())' as a specified value should serialize as 'calc(pow(sibling-index(), sibling-index()))'. assert_equals: 'calc(pow(sibling-index(), sibling-index()))' should round-trip exactly in specified values. expected "calc(pow(sibling-index(), sibling-index()))" but got "pow(sibling-index(), sibling-index())"
PASS 'calc(pow(sibling-index(), sibling-index())' as a computed value should serialize as '1'.
